Migrate gas/nonce fields to ulong
- Transaction: Nonce/GasLimit/SpentGas are ulong

- BlockHeader: GasLimit/GasUsed are ulong (Number stays long)

- TxReceipt: GasUsed/GasUsedTotal are ulong (BlockNumber stays long)

- Propagate chain-head gas limit through TxPool as ulong to reduce casts

+++ src/Nethermind/Nethermind.Facade/BlockchainBridge.cs
@@ -230,7 +230,7 @@ namespace Nethermind.Facade
             transaction.SenderAddress ??= Address.Zero;
 
             //Ignore nonce on all CallAndRestore calls
-            transaction.Nonce = components.StateReader.GetNonce(blockHeader, transaction.SenderAddress);
+            transaction.Nonce = (ulong)components.StateReader.GetNonce(blockHeader, transaction.SenderAddress);
 
             BlockHeader callHeader = treatBlockHeaderAsParentBlock
                 ? new(
